--- a/.flake8
+++ b/.flake8
@@ -1,3 +1,4 @@
 [flake8]
+extend-ignore = E203
 exclude = docs
 max-line-length = 188