--- a/src/bluetooth_data_tools/gap.pxd
+++ b/src/bluetooth_data_tools/gap.pxd
@@ -7,7 +7,6 @@ cdef object from_bytes
 cdef object from_bytes_little
 cdef object from_bytes_signed
 
-cdef object _cached_uint64_bytes_as_uuid
 cdef object _cached_uint16_bytes_as_uuid
 cdef object _cached_uint32_bytes_as_uuid
 cdef object _cached_uint128_bytes_as_uuid