This is ~42% speed up to the uncached parser

before

% python3 bench/parse_gap_tuple.py
Parsing 100000 bluetooth messages took 0.023806041994248517 seconds
Parsing 100000 bluetooth messages without caching took 1.3921346249990165 seconds

after

% python3 bench/parse_gap_tuple.py
Parsing 100000 bluetooth messages took 0.022713750004186295 seconds
Parsing 100000 bluetooth messages without caching took 0.8094419999979436 seconds